20230710 湖北朗志 修改ERP端订单回写采购单价和采购金额的数值精度

+++ src/main/java/hykx/ds/whkc/bean/ERPddmx.java
@@ -8,8 +8,8 @@ public class ERPddmx {
     private String batchnum;
     private String beizhu;
     private String cdmc;
-    private float cgdj;
-    private float cgje;
+    private double cgdj;
+    private double cgje;
     private float dj;
     private String dj_sn;
     private String djbh;
